pipe logs into log files as they run

+++ apps/scripts/loadgen-run.sh
@@ -28,22 +28,29 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
+if [[ -z $output_path ]]; then
+    echo "--output must be set"
+    exit 1
+fi
+
 container_name=hotel_client_bench
 network=local_hotel_network
 
 docker rm -f $container_name
 
+mkdir -p $output_path
+
 # Make sure that this is consistent with the network name created by docker compose
 docker run \
     --name $container_name \
     --network $network \
-    hotel_client_bench
+    hotel_client_bench \
+    &> $output_path/loadgen.log
 
 # this should be hard-coded in the docker entrypoint.sh
 container_trace_path="/tmp/masa-load-gen"
 
 # Copy traces from inside the container to outside
-mkdir -p $output_path
 docker cp $container_name:$container_trace_path $output_path
 # hacky way to make sure that the trace files are actually placed in $output_path
 mv $output_path/masa-load-gen/* $output_path
